refactor: move tracked category validation into the budgets service
